requirements.txt
selenium>=4.23.1
python-dotenv>=1.0.1
requests>=2.32.3
google-genai
